Adding the slot id exchange between xocl and zocl

+++ src/runtime_src/core/include/xgq_cmd_vmr.h
@@ -209,7 +209,7 @@ struct xgq_cmd_data_payload {
 	uint32_t flash_type:4;
 	uint32_t rsvd1:24;
 	uint32_t pad1;
-	uint64_t priv;
+	uint32_t priv;
 };
 
 enum xgq_cmd_flash_type {
